week5 assignment
