+++ static/css/attraction.css
@@ -1,11 +1,9 @@
 body{
   font-family: Noto Sans TC;
   margin :auto 0;
-  color: #666666;
-
-  
+  color: #666666;  
 }
-#navbar_container{
+/* #navbar_container{
   background-color: white;
   display: flex;
   justify-content: center;
@@ -35,16 +33,7 @@ body{
   text-align: end;
   margin: 10px;
   width: 53%;
-}
-#webFunction_book{
-  font-size: 16px;
-}
-#webFunction_signin{
-  font-size: 16px;
-}
-#webFunction_signout{
-  font-size: 16px;
-}
+} */
 .outside_container{
   width: 100%;
   display: flex;
@@ -174,20 +163,17 @@ body{
   align-items: center;
 
 }
-.book_trip_container > .book_time_container > .circle1 > .circle2{
+ #circle2{
   width: 18px;
   height: 18px;
   border-radius: 50%;
   background-color: #448899;
-  display: block;
 }
-.book_trip_container > .book_time_container > .circle3 > .circle4{
+ #circle4{
   width: 18px;
   height: 18px;
   border-radius: 50%;
   background-color: #448899;
-  display: none;
-
 }
 
 .book_trip_container > .book_money_container{
@@ -248,12 +234,12 @@ hr{
 }
 
 @media(min-width: 1200.1px){
-  .left{
+  /* .left{
     width: 80%;
   }
   .right{
     width: 20%;
-  }
+  } */
   .back_container{
     display: grid;
     grid-template-columns: 0.9fr 1.1fr;
@@ -282,12 +268,12 @@ hr{
 
 }
 @media(max-width: 1200px){
-  .left{
+  /* .left{
     width: 70%;
   }
   .right{
     width: 30%;
-  }
+  } */
   .back_container{
     display:grid ;
     grid-template-columns: 1.3fr 0.7fr;
@@ -309,12 +295,12 @@ hr{
 
 }
 @media(max-width: 900px){
-  .left{
+  /* .left{
     width: 47%;
   }
   .right{
     width: 53%;
-  }
+  } */
   .back_container{
     display: block;
   }  
@@ -341,131 +327,3 @@ hr{
   
 }
 
-/* 會員登入/註冊框 */
-#signin_box{
-  display: none;
-  margin: auto;
-  width: 340px;
-}
-#signup_box{
-  display: none;
-  margin: auto;
-  width: 340px;
-}
-.sign_top{
-  width: 340px;
-  height: 10px;
-  background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
-  border-radius: 6px 6px 0px 0px;
-}
-.sign{
-  width: 340px;
-  background-color: white;
-  border-radius: 0px 0px 6px 6px;
-
-}
-.sign_slogan_container{
-  display: grid;
-  grid-template-columns: 1.9fr 0.1fr;
-}
-.sign_slogan{
-  margin-top: 15px;
-  margin-bottom: 15px;
-  margin-left:100px;
-  font-size: 24px;
-  font-weight: 700;
-  color: #666666;
-}
-.sign_slogan_container > img{
-  width: 16px;
-  height: 16px;
-  margin:17px;
-}
-.sign > input{
-  width: 292px;
-  height: 41px;
-  margin: 0px 15px 10px 15px;
-  font-size: 16px;
-  font-weight: 400;
-  padding-left: 10px;
-  border: 1px solid;
-  border-color: #CCCCCC;
-  border-radius: 5px;
-}
-.sign > button{
-  width: 310px;
-  height: 47px;
-  margin: 0px 15px 10px 15px;
-  background-color: #448899;
-  border: 0;
-  border-radius: 5px;
-  color :white;
-  font-size: 19px;
-}
-#signup_message{
-  font-size: 16px;
-  color: rgb(202, 27, 27);
-  width: 100%;
-  text-align: center;
-  padding-bottom: 5px;
-}
-#signin_message{
-  font-size: 16px;
-  color: rgb(202, 27, 27);
-  width: 100%;
-  text-align: center;
-  padding-bottom: 5px;
-}
-.a {
-  width: 100%;
-  padding-bottom: 15px;
-  display: flex;
-  justify-content: center;
-}
-.a > button{
-  font-size: 16px;
-  font-weight: 400;
-  color: #666666;
-  border:0;
-  background-color: white;
-
-}
-
-#modal{
-  display: none;
-  position: fixed;
-  z-index: 2;
-  padding-top: 80px;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow:auto;
-  background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
-}
-
-
-.close_button:hover {
-  cursor: pointer;
-}
-#backto_signup:hover{
-  cursor: pointer;
-}
-#backto_signin:hover{
-  cursor: pointer;
-}
-#signin_button:hover{
-  cursor: pointer;
-}
-#signup_button:hover{
-  cursor: pointer;
-}
-#webFunction_book:hover{
-  cursor: pointer;
-}
-#webFunction_signin:hover{
-  cursor: pointer;
-}
-#webFunction_signout:hover{
-  cursor: pointer;
-}
