+++ static/css/index.css
@@ -1,0 +1,163 @@
+body{
+  font-family: Noto Sans TC;
+  margin :auto 0;
+}
+.container{
+  width: 100vw;
+  background: linear-gradient(135deg, rgba(170,221,238) 0%, rgba(102,170,187) 100%);
+  margin-top: 30px;
+  display: flex;
+  justify-content: center;
+}
+
+.backgroundPictureContainer{
+  width: 100vw;
+  background: url(../images/freecourt.png) right;
+  background-repeat: no-repeat;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: #F8F8F8;
+
+}
+.sloganContainer{
+  width: 1200px;
+}
+#slogan1{
+  margin: 10px 0px 0px 10px;
+  font-weight: bold;
+  font-size: 28px;
+}
+#slogan2{
+  margin: 10px 0px 15px 10px;
+  font-size: 16px;
+}
+.searchContainer{
+  display: flex;
+  margin: 0px 0px 10px 10px;
+}
+input{
+  border: 0;
+  height: 44px;
+  border-radius: 5px 0px 0px 5px;
+  padding-left: 10px;
+}
+#searchButton{
+  border: 0;
+  width: 60px;
+  height: 46px;
+  background: url(../images/icon_search.png)center no-repeat, #448899;
+  border-radius: 0px 5px 5px 0px;
+}
+.attractionsContainer{
+  display: flex;
+  flex-wrap: wrap;                    /*使下一項目排序若超過容器寬度則自動換行*/
+  justify-content: space-evenly;      /*除了margin外，將剩餘空間平均分配*/
+  margin:auto;
+  padding-top: 50px;
+  
+}
+.attractionBox{
+  width: 270px; 
+  border: 1px solid #E8E8E8;
+  border-radius: 5px;
+  margin: 10px;
+}
+img{
+  width: 268px;                     /*將圖片寬度固定*/
+  height: 160px;                  /*將圖片高度固定*/
+  object-fit: cover;                /*隨框架縮放並維持圖片比例*/
+}
+.attractionName{
+  height: 32px;
+  font-size: 16px;
+  font-weight: bold;
+  padding: 10px 0px 0px 10px;
+  color: #757575;
+}
+.attractionInformation{
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  height: 20px;
+  align-items: center;
+  margin: 10px;
+}
+.attractionMrt{ 
+  font-size: 16px;
+  color: #757575;
+} 
+.attractionCategory{
+  text-align: right;
+  font-size: 16px;
+  color: #757575;
+}
+.footerContainer{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #757575;
+  color: white;
+  height: 104px;
+}
+
+
+@media(min-width: 1200.1px){
+  .backgroundPictureContainer{
+    background-position: 80% 100%;
+    height: 320px;
+  }
+  input{
+    width: 400px;
+  }
+  .attractionsContainer{
+    width: 1200px;
+  }
+}
+@media(max-width: 1200px){
+  .backgroundPictureContainer{
+    background-position: 110% 100%;
+    background-size: 65%;
+    height: 320px;
+  }
+  input{
+    width: 400px;
+  }
+  .attractionsContainer{
+    width: 90%;
+  }
+  .attractionBox{
+    width: 45%;
+  }
+  img{
+    width: 100%;
+  }
+}
+
+@media(max-width: 600px){
+  #slogan1{
+    margin: 10px 0px 0px 15px;
+  }
+  #slogan2{
+    margin: 10px 0px 15px 15px;
+  }
+  .searchContainer{
+    margin: 0px 0px 10px 15px;
+  }
+  .backgroundPictureContainer{
+    background-position: 360% 100%;
+    background-size: 90%;
+    height: 280px;
+  }
+  input{
+    width: 230px;
+  }
+  .attractionBox{
+    width: 90%;
+    margin: 10px 0px 10px 0px;
+  }
+  img{
+    width: 100%;
+    height: 60vw;
+  }
+}
+
