+++ static/css/share.css
@@ -1,0 +1,219 @@
+body{
+  min-height: 100vh;
+  position: relative;
+}
+.wrapper{
+  padding-bottom: 104px;
+}
+/* navbar */
+#navbar_container{
+  background-color: white;
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 1;
+  border-bottom: #E8E8E8 1px solid;
+}
+.navbar{
+  padding-top: 10px;
+  display: flex;
+}
+.left{
+  width: 50%;
+}
+#trip_name{
+  width:155px;
+  text-decoration: none;
+  font-weight: bold;
+  font-size: 30px;
+  color: #448899;
+}
+.right{
+  width: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  color: #666666;
+}
+#webFunction_book{
+  margin:  0px 10px 0px 10px;
+  font-size: 16px;
+}
+#webFunction_signin{
+  margin:  0px 10px 0px 10px;
+  font-size: 16px;
+  display: block;
+}#webFunction_signout{
+  margin:  0px 10px 0px 10px;
+  font-size: 16px;
+}
+@media(min-width: 1200.1px){
+  .navbar{
+    width: 1200px;
+  }
+}
+@media(max-width: 1200px){
+  .navbar{
+    width: 1200px;
+  }
+  .left{
+    margin-left: 15px;
+  }
+}
+@media(max-width: 600px){
+  .left{
+    margin-left: 15px;
+    width: 45%;
+  }
+  .right{
+    width: 55%;
+  }
+}
+/*footer */
+.footerContainer{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #757575;
+  color: white;
+  width: 100%;
+  height: 104px;
+  position: absolute;
+  bottom: 0;
+}
+
+/* 會員登入/註冊框 */
+#modal{
+  display: none;
+  position: fixed;
+  z-index: 2;
+  padding-top: 80px;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow:auto;
+  background-color: rgba(0,0,0,0.4); 
+}
+
+#signin_box{
+  display: none;
+  margin: auto;
+  width: 340px;
+}
+#signup_box{
+  display: none;
+  margin: auto;
+  width: 340px;
+}
+.sign_top{
+  width: 340px;
+  height: 10px;
+  background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
+  border-radius: 6px 6px 0px 0px;
+}
+.sign{
+  width: 340px;
+  background-color: white;
+  border-radius: 0px 0px 6px 6px;
+
+}
+.sign_slogan_container{
+  display: grid;
+  grid-template-columns: 1.9fr 0.1fr;
+}
+.sign_slogan{
+  margin-top: 15px;
+  margin-bottom: 15px;
+  margin-left:100px;
+  font-size: 24px;
+  font-weight: 700;
+  color: #666666;
+}
+.sign_slogan_container > img{
+  width: 16px;
+  height: 16px;
+  margin:17px;
+}
+.sign > input{
+  width: 292px;
+  height: 41px;
+  margin: 0px 15px 10px 15px;
+  font-size: 16px;
+  font-weight: 400;
+  padding-left: 10px;
+  border: 1px solid;
+  border-color: #CCCCCC;
+  border-radius: 5px;
+}
+.sign > button{
+  width: 310px;
+  height: 47px;
+  margin: 0px 15px 10px 15px;
+  background-color: #448899;
+  border: 0;
+  border-radius: 5px;
+  color :white;
+  font-size: 19px;
+}
+#signup_message{
+  font-size: 16px;
+  color: rgb(202, 27, 27);
+  width: 100%;
+  text-align: center;
+  padding-bottom: 5px;
+}
+#signin_message{
+  font-size: 16px;
+  color: rgb(202, 27, 27);
+  width: 100%;
+  text-align: center;
+  padding-bottom: 5px;
+}
+.a {
+  width: 100%;
+  padding-bottom: 15px;
+  display: flex;
+  justify-content: center;
+}
+.a > button{
+  font-size: 16px;
+  font-weight: 400;
+  color: #666666;
+  border:0;
+  background-color: white;
+
+}
+
+
+
+#trip_name:hover{
+  cursor: pointer;
+}
+.close_button:hover {
+  cursor: pointer;
+}
+#backto_signup:hover{
+  cursor: pointer;
+}
+#backto_signin:hover{
+  cursor: pointer;
+}
+#signin_button:hover{
+  cursor: pointer;
+}
+#signup_button:hover{
+  cursor: pointer;
+}
+#webFunction_book:hover{
+  cursor: pointer;
+}
+#webFunction_signin:hover{
+  cursor: pointer;
+}
+#webFunction_signout:hover{
+  cursor: pointer;
+}
+
