+++ static/images/icon_search.png
