--- static/index.css
@@ -1,347 +1,0 @@
-body{
-  font-family: Noto Sans TC;
-  margin :auto 0;
-  
-}
-#navbarContainer{
-  background-color: white;
-  display: flex;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  width: 100%;
-}
-.navbar{
-  padding-top: 10px;
-  display: flex;
-}
-.left{
-  font-weight: bold;
-  font-size: 30px;
-  width: 50%;
-  color: #448899;
-  
-}
-.right{
-  width: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  color: #666666;
-}
-#webFunction_book{
-  margin:  0px 10px 0px 10px;
-  font-size: 16px;
-}
-#webFunction_signin{
-  margin:  0px 10px 0px 10px;
-  font-size: 16px;
-  display: block;
-}#webFunction_signout{
-  margin:  0px 10px 0px 10px;
-  font-size: 16px;
-  /* display: block; */
-}
-.container{
-  width: 100vw;
-  background: linear-gradient(135deg, rgba(170,221,238) 0%, rgba(102,170,187) 100%);
-  margin-top: 30px;
-  display: flex;
-  justify-content: center;
-}
-
-.backgroundPictureContainer{
-  width: 100vw;
-  background: url(welcome\ 1.png) right;
-  background-repeat: no-repeat;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #F8F8F8;
-
-}
-.sloganContainer{
-  width: 1200px;
-}
-#slogan1{
-  margin: 10px 0px 0px 10px;
-  font-weight: bold;
-  font-size: 28px;
-}
-#slogan2{
-  margin: 10px 0px 15px 10px;
-  font-size: 16px;
-}
-.searchContainer{
-  display: flex;
-  margin: 0px 0px 10px 10px;
-}
-input{
-  border: 0;
-  height: 44px;
-  border-radius: 5px 0px 0px 5px;
-  padding-left: 10px;
-}
-#searchButton{
-  border: 0;
-  width: 60px;
-  height: 46px;
-  background: url(icon_search.png)center no-repeat, #448899;
-  border-radius: 0px 5px 5px 0px;
-}
-.attractionsContainer{
-  display: flex;
-  flex-wrap: wrap;                    /*使下一項目排序若超過容器寬度則自動換行*/
-  justify-content: space-evenly;      /*除了margin外，將剩餘空間平均分配*/
-  margin:auto;
-  padding-top: 50px;
-  
-}
-.attractionBox{
-  width: 270px; 
-  border: 1px solid #E8E8E8;
-  border-radius: 5px;
-  margin: 10px;
-}
-img{
-  width: 268px;                     /*將圖片寬度固定*/
-  height: 160px;                  /*將圖片高度固定*/
-  object-fit: cover;                /*隨框架縮放並維持圖片比例*/
-}
-.attractionName{
-  height: 32px;
-  font-size: 16px;
-  font-weight: bold;
-  padding: 10px 0px 0px 10px;
-  color: #757575;
-}
-.attractionInformation{
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  height: 20px;
-  align-items: center;
-  margin: 10px;
-}
-.attractionMrt{ 
-  font-size: 16px;
-  color: #757575;
-} 
-.attractionCategory{
-  text-align: right;
-  font-size: 16px;
-  color: #757575;
-}
-.footerContainer{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #757575;
-  color: white;
-  height: 104px;
-}
-
-
-@media(min-width: 1200.1px){
-  .navbar{
-    width: 1200px;
-  }
-  .backgroundPictureContainer{
-    background-position: 80% 100%;
-    height: 320px;
-  }
-  input{
-    width: 400px;
-  }
-  .attractionsContainer{
-    width: 1200px;
-  }
-}
-@media(max-width: 1200px){
-  .navbar{
-    width: 1200px;
-  }
-  .left{
-    margin-left: 15px;
-  }
-  .backgroundPictureContainer{
-    background-position: 110% 100%;
-    background-size: 65%;
-    height: 320px;
-  }
-  input{
-    width: 400px;
-  }
-  .attractionsContainer{
-    width: 90%;
-  }
-  .attractionBox{
-    width: 45%;
-  }
-  img{
-    width: 100%;
-  }
-}
-
-@media(max-width: 600px){
-  .left{
-    margin-left: 15px;
-    width: 45%;
-  }
-  .right{
-    width: 55%;
-  }
-  #slogan1{
-    margin: 10px 0px 0px 15px;
-  }
-  #slogan2{
-    margin: 10px 0px 15px 15px;
-  }
-  .searchContainer{
-    margin: 0px 0px 10px 15px;
-  }
-  .backgroundPictureContainer{
-    background-position: 360% 100%;
-    background-size: 90%;
-    height: 280px;
-  }
-  input{
-    width: 230px;
-  }
-  .attractionBox{
-    width: 90%;
-    margin: 10px 0px 10px 0px;
-  }
-  img{
-    width: 100%;
-    height: 60vw;
-  }
-}
-/* 會員登入/註冊框 */
-#signin_box{
-  display: none;
-  margin: auto;
-  width: 340px;
-}
-#signup_box{
-  display: none;
-  margin: auto;
-  width: 340px;
-}
-.sign_top{
-  width: 340px;
-  height: 10px;
-  background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
-  border-radius: 6px 6px 0px 0px;
-}
-.sign{
-  width: 340px;
-  background-color: white;
-  border-radius: 0px 0px 6px 6px;
-
-}
-.sign_slogan_container{
-  display: grid;
-  grid-template-columns: 1.9fr 0.1fr;
-}
-.sign_slogan{
-  margin-top: 15px;
-  margin-bottom: 15px;
-  margin-left:100px;
-  font-size: 24px;
-  font-weight: 700;
-  color: #666666;
-}
-.sign_slogan_container > img{
-  width: 16px;
-  height: 16px;
-  margin:17px;
-}
-.sign > input{
-  width: 292px;
-  height: 41px;
-  margin: 0px 15px 10px 15px;
-  font-size: 16px;
-  font-weight: 400;
-  padding-left: 10px;
-  border: 1px solid;
-  border-color: #CCCCCC;
-  border-radius: 5px;
-}
-.sign > button{
-  width: 310px;
-  height: 47px;
-  margin: 0px 15px 10px 15px;
-  background-color: #448899;
-  border: 0;
-  border-radius: 5px;
-  color :white;
-  font-size: 19px;
-}
-#signup_message{
-  font-size: 16px;
-  color: rgb(202, 27, 27);
-  width: 100%;
-  text-align: center;
-  padding-bottom: 5px;
-}
-#signin_message{
-  font-size: 16px;
-  color: rgb(202, 27, 27);
-  width: 100%;
-  text-align: center;
-  padding-bottom: 5px;
-}
-.a {
-  width: 100%;
-  padding-bottom: 15px;
-  display: flex;
-  justify-content: center;
-}
-.a > button{
-  font-size: 16px;
-  font-weight: 400;
-  color: #666666;
-  border:0;
-  background-color: white;
-
-}
-
-#modal{
-  display: none;
-  position: fixed;
-  z-index: 2;
-  padding-top: 80px;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow:auto;
-  background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
-}
-
-
-.close_button:hover {
-  cursor: pointer;
-}
-#backto_signup:hover{
-  cursor: pointer;
-}
-#backto_signin:hover{
-  cursor: pointer;
-}
-#signin_button:hover{
-  cursor: pointer;
-}
-#signup_button:hover{
-  cursor: pointer;
-}
-#webFunction_book:hover{
-  cursor: pointer;
-}
-#webFunction_signin:hover{
-  cursor: pointer;
-}
-#webFunction_signout:hover{
-  cursor: pointer;
-}
-
