+++ templates/attraction.html
@@ -5,111 +5,73 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title></title>
-    <link rel="stylesheet" type="text/css" href="{{url_for('static', filename= 'attraction.css') }}" />
+    <title>attractions</title>
+    <link rel="stylesheet" type="text/css" href= "../static/css/share.css" />
+    <link rel="stylesheet" type="text/css" href=" ../static/css/attraction.css"/>
 </head>
 <body>
-  <div id= "navbar_container">
-    <nav class="navbar">
-      <div class= "left">台北一日遊</div>
-      <div class= "right">
-        <div id= "webFunction_book">預定行程</div>
-        <div id= "webFunction_signin">登入/註冊</div>
-        <divc id= "webFunction_signout">登出系統</div>
-      </div>
-    </nav>
-  </div>
-  <div class= "outside_container">
-    <div class= "back_container">
-      <div class= "attractions_container">
-        <div class= "attraction_image">
-          <div class= "arrow_button_container">
-            <img src= "../static/btn_leftArrow.png" id= "arrow_button_left">
-            <img src= "../static/btn_rightArrow.png" id= "arrow_button_right">
+  <div class = "wrapper">
+    <!--navbar-->
+    {% include "/share-html/navbar.html" %}
+    <div class= "outside_container">
+      <div class= "back_container">
+        <div class= "attractions_container">
+          <div class= "attraction_image">
+            <div class= "arrow_button_container">
+              <img src= "../static/images/btn_leftArrow.png" id= "arrow_button_left">
+              <img src= "../static/images/btn_rightArrow.png" id= "arrow_button_right">
+            </div>
           </div>
         </div>
-      </div>
-      <div class= "outside_book_container">
-        <div class= "book_container">
-          <div class= "attraction_information_container"></div>
-          <div class= "book_trip_container">
-            <p class= "book_information1">訂購導覽行程</p>
-            <p class= "book_information2">以此景點為中心的一日行程，帶您探索城市角落故事</p>
-            <div class= "book_date_container">
-              <p class= "book_date1">選擇日期：</p>
-              <input type= "date"></input>
+        <div class= "outside_book_container">
+          <div class= "book_container">
+            <div class= "attraction_information_container"></div>
+            <div class= "book_trip_container">
+              <p class= "book_information1">訂購導覽行程</p>
+              <p class= "book_information2">以此景點為中心的一日行程，帶您探索城市角落故事</p>
+              <div class= "book_date_container">
+                <p class= "book_date1">選擇日期：</p>
+                <input type= "date" id = "book_date"></input>
+              </div>
+              <div class= "book_time_container">
+                <p class= "book_time">選擇時間：</p>
+                <div class= "circle1"><div id= "circle2" style="display: block"></div></div>
+                <label >上半天</label>     
+                <div class= "circle3"><div id= "circle4" style="display: none"></div></div>
+                <label >下半天</label>       
+              </div>
+              <div class= "book_money_container">
+                <p class= "book_money">導覽費用：
+                  <div class= "money"> 新台幣2000元</div>
+                </p>
+              </div>
+              <button id= "book_button" type = "button">開始預訂行程</button>
             </div>
-            <div class= "book_time_container">
-              <p class= "book_time">選擇時間：</p>
-              <div class= "circle1"><div class= "circle2"></div></div>
-              <label >上半天</label>     
-              <div class= "circle3"><div class= "circle4"></div></div>
-              <label >下半天</label>       
-            </div>
-            <div class= "book_money_container">
-              <p class= "book_money">導覽費用：
-                <div class= "money"> 新台幣2000元</div>
-              </p>
-            </div>
-            <button id= "book_button">開始預訂行程</button>
           </div>
         </div>
       </div>
     </div>
-  </div>
-  <div id= "hr_container">
-    <hr>
-  </div>
-  <div class= "attraction_detail_container">
-    <div class= "attraction_detail">
-      <div class= "attraction_description"></div>
-      <div class= "attraction_address_title">景點地址：</div>
-      <div id= "attraction_address"></div>
-      <div class= "attraction_transport_title">交通方式：</div>
-      <div id= "attraction_transport"></div>
+    <div id= "hr_container">
+      <hr>
     </div>
-  </div>
-  <!--會員登入/註冊框-->
-  <div id = "modal">
-    <div id = "signin_box">
-      <div class = "sign_top"></div>
-      <div class = "sign">
-        <div class = "sign_slogan_container" >
-          <p class = "sign_slogan">登入會員帳號</p>
-          <img src = "../static/icon_close.png" class = "close_button">
-        </div>
-        <input id = "signin_email" type = "email" placeholder = "輸入電子信箱">
-        <input id = "signin_password" type = "password" placeholder = "輸入密碼">
-        <button id = "signin_button">登入帳戶</button> 
-        <div id = "signin_message"></div>
-        <div class = "a">
-          <button id = "backto_signup">還沒有帳戶？點此註冊</a>
-        </div>
+    <div class= "attraction_detail_container">
+      <div class= "attraction_detail">
+        <div class= "attraction_description"></div>
+        <div class= "attraction_address_title">景點地址：</div>
+        <div id= "attraction_address"></div>
+        <div class= "attraction_transport_title">交通方式：</div>
+        <div id= "attraction_transport"></div>
       </div>
     </div>
-    <div id = "signup_box">
-      <div class = "sign_top"></div>
-      <div class = "sign">
-        <div class = "sign_slogan_container" >
-          <p class = "sign_slogan">註冊會員帳號</p>
-          <img src = "../static/icon_close.png" class = "close_button">
-        </div>
-        <input id = "signup_name" type = "text" placeholder = "輸入姓名">
-        <input id = "signup_email" type = "email" placeholder = "輸入電子郵件">
-        <input id = "signup_password" type = "password" placeholder = "輸入密碼">
-        <button id = "signup_button">註冊新帳戶</button> 
-        <div id = "signup_message"></div>
-        <div class = "a">
-          <button id = "backto_signin">已有帳戶了？點此登入</a>
-        </div>
-      </div>
-    </div>
+    <!--會員登入/註冊框-->
+    {% include "/share-html/member.html" %}
+    <!--footer-->
+    {% include "share-html/footer.html" %}
   </div>
-  <div class= "footerContainer">
-    <footer class= "footer">COPYRIGHT © 2021 台北一日遊</footer>
-  </div>
-  
-<script src= "{{ url_for("static", filename= "attraction.js") }}"  type= "text/javascript"></script>
+  <script src= "../static/js/member.js"  type= "text/javascript"></script>
+  <script src= "../static/js/attraction.js"  type= "text/javascript"></script>
+
+
 
 </body>
 </html>
