+++ templates/index.html
@@ -6,79 +6,39 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>首頁</title>
-    <link rel="stylesheet" type="text/css" href="{{url_for('static', filename= 'index.css') }}" />
+    <link rel="stylesheet" type="text/css" href= "../static/css/share.css" />
+    <link rel="stylesheet" type="text/css" href= "../static/css/index.css" />
 </head>
 <body>
-  <div id= "navbarContainer">
-    <nav class="navbar">
-      <div class= "left">台北一日遊</div>
-      <div class= "right">
-        <div id= "webFunction_book">預定行程</div>
-        <div id= "webFunction_signin">登入/註冊</div>
-        <divc id= "webFunction_signout">登出系統</div>
-      </div>
-    </nav>
-  </div>
-  <br>
-  <div class= "container">
-    <div class= "backgroundPictureContainer">
-      <div class= "sloganContainer">
-        <div id= "slogan1">輕鬆享受台北一日悠閒</div>
-        
-        <div id= "slogan2">探索每個角落，體驗城市的深度旅遊行程</div>
-        
-        <div class="searchContainer">
+  <div class = "wrapper">
+    <!--navbar-->
+    {% include "/share-html/navbar.html" %}
+    <br>
+    <div class= "container">
+      <div class= "backgroundPictureContainer">
+        <div class= "sloganContainer">
+          <div id= "slogan1">輕鬆享受台北一日悠閒</div>
           
-            <input id= "inputKeyword" type= "text" placeholder= "輸入景點名稱查詢" >
-            <button id= "searchButton" type= "submit" ></button>
+          <div id= "slogan2">探索每個角落，體驗城市的深度旅遊行程</div>
           
+          <div class="searchContainer">
+            
+              <input id= "inputKeyword" type= "text" placeholder= "輸入景點名稱查詢" >
+              <button id= "searchButton" type= "submit" ></button>
+            
+          </div>
         </div>
       </div>
     </div>
-  </div>
-  <div class="attractionsContainer">
-  </div>
-  <!--會員登入/註冊框-->
-  <div id = "modal">
-    <div id = "signin_box">
-      <div class = "sign_top"></div>
-      <div class = "sign">
-        <div class = "sign_slogan_container" >
-          <p class = "sign_slogan">登入會員帳號</p>
-          <img src = "../static/icon_close.png" class = "close_button">
-        </div>
-        <input id = "signin_email" type = "email" placeholder = "輸入電子信箱">
-        <input id = "signin_password" type = "password" placeholder = "輸入密碼">
-        <button id = "signin_button">登入帳戶</button> 
-        <div id = "signin_message"></div>
-        <div class = "a">
-          <button id = "backto_signup">還沒有帳戶？點此註冊</a>
-        </div>
-      </div>
+    <div class="attractionsContainer">
     </div>
-    <div id = "signup_box">
-      <div class = "sign_top"></div>
-      <div class = "sign">
-        <div class = "sign_slogan_container" >
-          <p class = "sign_slogan">註冊會員帳號</p>
-          <img src = "../static/icon_close.png" class = "close_button">
-        </div>
-        <input id = "signup_name" type = "text" placeholder = "輸入姓名">
-        <input id = "signup_email" type = "email" placeholder = "輸入電子郵件">
-        <input id = "signup_password" type = "password" placeholder = "輸入密碼">
-        <button id = "signup_button">註冊新帳戶</button> 
-        <div id = "signup_message"></div>
-        <div class = "a">
-          <button id = "backto_signin">已有帳戶了？點此登入</a>
-        </div>
-      </div>
-    </div>
+    <!--會員登入/註冊框-->
+    {% include "/share-html/member.html" %}
+    <!--footer-->
+    {% include "share-html/footer.html" %}
   </div>
-  <div class= "footerContainer">
-    <footer class= "footer">COPYRIGHT © 2021 台北一日遊</footer>
-  </div>
-
-  <script src= "{{ url_for("static", filename= "index.js") }}"  type= "text/javascript"></script>
+  <script src= "../static/js/member.js"  type= "text/javascript"></script>
+  <script src= "../static/js/index.js"  type= "text/javascript"></script>
 </body>
 </html>
