+++ templates/share-html/footer.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+</head>
+<body>
+  <div class= "footerContainer">
+    <footer class= "footer">COPYRIGHT © 2021 台北一日遊</footer>
+  </div>
+</body>
+</html>
