+++ templates/share-html/member.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+</head>
+<body>
+  <div id = "modal">
+    <div id = "signin_box">
+      <div class = "sign_top"></div>
+      <div class = "sign">
+        <div class = "sign_slogan_container" >
+          <p class = "sign_slogan">登入會員帳號</p>
+          <img src = "../static/images/icon_close.png" class = "close_button">
+        </div>
+        <input id = "signin_email" type = "email" placeholder = "輸入電子信箱">
+        <input id = "signin_password" type = "password" placeholder = "輸入密碼">
+        <button id = "signin_button">登入帳戶</button> 
+        <div id = "signin_message"></div>
+        <div class = "a">
+          <button id = "backto_signup">還沒有帳戶？點此註冊</a>
+        </div>
+      </div>
+    </div>
+    <div id = "signup_box">
+      <div class = "sign_top"></div>
+      <div class = "sign">
+        <div class = "sign_slogan_container" >
+          <p class = "sign_slogan">註冊會員帳號</p>
+          <img src = "../static/images/icon_close.png" class = "close_button">
+        </div>
+        <input id = "signup_name" type = "text" placeholder = "輸入姓名">
+        <input id = "signup_email" type = "email" placeholder = "輸入電子郵件">
+        <input id = "signup_password" type = "password" placeholder = "輸入密碼">
+        <button id = "signup_button">註冊新帳戶</button> 
+        <div id = "signup_message"></div>
+        <div class = "a">
+          <button id = "backto_signin">已有帳戶了？點此登入</a>
+        </div>
+      </div>
+    </div>
+  </div>
+</body>
+</html>
