+++ templates/share-html/navbar.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+</head>
+<body>
+  <div id= "navbar_container">
+    <nav class="navbar">
+      <div class= "left">
+        <a id = "trip_name" href= "/">台北一日遊</a>
+      </div>
+      <div class= "right">
+        <div id= "webFunction_book">預定行程</div>
+        <div id= "webFunction_signin">登入/註冊</div>
+        <div id= "webFunction_signout">登出系統</div>
+      </div>
+    </nav>
+  </div>
+</body>
+</html>
